- Added descriptions to some debug flags. - Removed "unused" debug flags. - Got rid of some compiler warnings in OMI.
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@16190 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/interactive/omi_Calculation.cpp
@@ -205,7 +205,7 @@ static void printSSDCalculation(long nStates, long nAlgebraic, long nParameters)
  */
 THREAD_RET_TYPE threadSimulationCalculation(THREAD_PARAM_TYPE lpParam)
 {
-  int retValue = -1; //Not used yet
+  long retValue = -1; //Not used yet
 
   if (debugCalculation) {
     cout << "Calculation:\tMessage: Calculation Thread Start*****" << endl;
